requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
click==8.1.7
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
playwright==1.41.1
pydantic==2.6.1
pydantic_core==2.16.2
pyee==11.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
